IntellBot/code/requirements.txt
langchain==0.0.184
# PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
# openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
chromadb==0.4.22
# uncomment to use huggingface llms
huggingface-hub==0.14.1
langchain_community==0.0.26
# uncomment to use instructor embeddings
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2






# aiohttp==3.8.4
# aiosignal==1.3.1
# altair==5.0.1
# async-timeout==4.0.2
# attrs==23.1.0
# blinker==1.6.2
# cachetools==5.3.1
# certifi==2023.5.7
# charset-normalizer==3.1.0
# click==8.1.3
# colorama==0.4.6
# dataclasses-json==0.5.7
# decorator==5.1.1
# faiss-cpu==1.7.4
# frozenlist==1.3.3
# gitdb==4.0.10
# GitPython==3.1.31
# greenlet==2.0.2
# idna==3.4
# importlib-metadata==6.6.0
# Jinja2==3.1.2
# joblib==1.2.0
# jsonschema==4.17.3
# langchain==0.0.189
# markdown-it-py==2.2.0
# MarkupSafe==2.1.3
# marshmallow==3.19.0
# marshmallow-enum==1.5.1
# mdurl==0.1.2
# multidict==6.0.4
# mypy-extensions==1.0.0
# nltk==3.8.1
# numexpr==2.8.4
# numpy==1.24.3
# openai==0.27.7
# openapi-schema-pydantic==1.2.4
# packaging==23.1
# pandas==2.0.2
# Pillow==9.5.0
# plotly==5.14.1
# protobuf==4.23.2
# pyarrow==12.0.0
# pydantic==1.10.8
# pydeck==0.8.1b0
# Pygments==2.15.1
# Pympler==1.0.1
# PyPDF2==3.0.1
# pyrsistent==0.19.3
# python-dateutil==2.8.2
# python-dotenv==1.0.0
# pytz==2023.3
# pytz-deprecation-shim==0.1.0.post0
# PyYAML==6.0
# regex==2023.6.3
# requests==2.31.0
# rich==13.4.1
# six==1.16.0
# smmap==5.0.0
# SQLAlchemy==2.0.15
# streamlit==1.23.1
# tenacity==8.2.2
# tiktoken==0.4.0
# toml==0.10.2
# toolz==0.12.0
# tornado==6.3.2
# tqdm==4.65.0
# typing-inspect==0.9.0
# typing_extensions==4.6.3
# tzdata==2023.3
# tzlocal==4.3
# urllib3==2.0.2
# validators==0.20.0
# watchdog==3.0.0
# Werkzeug==2.3.4
# yarl==1.9.2
# zipp==3.15.0

# css='''
# <style>
# .chat-container {
#   font-family: sans-serif;
#   display: flex;
#   flex-direction: column;
#   height: 500px;
#   overflow-y: scroll;
#   padding: 1rem;
# }

# .chat-message {
#   display: flex;
#   margin-bottom: 1rem;
#   padding: 1rem;
#   border-radius: 0.5rem;
# }

# .chat-message.user {
#   background-color: #f5f5f5;
#   align-self: flex-end;
# }
# .chat-message.bot {
#   background-color: #ddd;
#   align-self: flex-start;
# }
# .chat-message .avatar {
#   width: 50px;
#   height: 50px;
#   border-radius: 50%;
#   margin-right: 1rem;
#   margin-left: 1rem;
# }

# .chat-message .avatar img {
#   max-width: 100%;
#   max-height: 100%;
#   object-fit: cover;
# }

# .chat-message .message {
#   flex: 1;
#   padding: 0.5rem;
#   color: #000;
#   font-size: 16px;
# }

# .chat-message .timestamp {
#   font-size: 0.8rem;
#   color: #999;
#   margin-top: 0.5rem;
# }

# .chat-input {
#   width: 100%;
#   padding: 1rem;
#   border: 1px solid #ddd;
#   border-radius: 0.5rem;
#   margin-top: 1rem;
# }
# .chat-send-button {
#   background-color: #4CAF50;
#   color: white;
#   padding: 10px 20px;
#   border: none;
#   border-radius: 4px;
#   cursor: pointer;
# }

# .chat-send-button:hover {
#   background-color: #45a049;
# }
# </style>
# '''

# bot_template = '''
# <div class="chat-message bot">
#     <div class="avatar">
#         <img src="https://i.ibb.co/X8Y5Rpg/robo.webp" alt="robo" border="0" />
#     </div>
#     <div class="message">{{MSG}}</div>
# </div>
# '''

# user_template = '''
# <div class="chat-message user">
#     <div class="avatar">
#         <img src="https://i.ibb.co/58V5f71/user.png" alt="user" border="0" />
#     </div>    
#     <div class="message">{{MSG}}</div>
# </div>
# '''
